requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
flake8==3.9.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1